third_party/WebKit/LayoutTests/fast/workers/worker-onerror-07.html

<!DOCTYPE html>
<html>
<head>
    <script>
        window.isOnErrorTest = true;
    </script>
    <script src="../../resources/js-test.js"></script>
    <script src="resources/onerror-test.js"></script>
</head>
<body>
    <!-- This script's body will be used to build a Blob URL to use as a Worker. -->
    <script id="workerCode" type="text/plain">
        onerror = function(message, filename, lineno, colno, error) {
            postMessage({
                message: message,
                filename: filename,
                lineno: lineno,
                colno: colno
            });
            setTimeout(function () { postMessage({ done: true }); }, 0);
            return true;
        }

        foo.bar = 0;
    </script>
    <script>
        description("Handling an error in the worker should suppress the page-level onerror handler.");

        checkPostMessage({
            message: "Uncaught ReferenceError: foo is not defined",
            filename: "[blob: URL]",
            lineno: 13,
            colno: 9,
        });

        errorEventHandlerShouldNotFire();
    </script>
</body>
</html>
